Fix Timeout Errors; Expand Logging Information
This commit logs an error to the output rather than crashing on socket error. To avoid timeout errors the server has been set to blocking mode rather than timeout mode.

The logging system has been expanded to include the side logging the message (client|server) and the time that the program logged the message.